backend/requirements.txt
fastapi
uvicorn[standard]

python-jose

passlib[bcrypt]==1.7.4
bcrypt==3.2.2

python-multipart
python-dotenv

google-auth
requests

sqlalchemy
alembic
